Move bep3 msg server to keeper
